♻️ refactor: Move `Footer` to a separate component and update `Header` import path
- Extracted `Footer` into a new component under `layout` for better modularity.
- Updated `Header` import path to align with the new directory structure.
- Adjusted `PermissionStep` layout by adding `mt-20` class to the section for consistent spacing.
