+++ app/components/layout/Footer.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const Footer = () => {
+    const currentYear = new Date().getFullYear();
+    return (
+        <footer className="py-4 text-center text-sm text-gray-500">
+            ©{currentYear}
+        </footer>
+    )
+}
+
+export default Footer;
