+++ app/layout.tsx
@@ -1,6 +1,7 @@
 import './globals.css'
-import Header from '@/components/Header'
+import Header from '@/components/layout/Header'
 import React from "react";
+import Footer from "@/components/layout/Footer";
 
 export const metadata = {
     title: 'Exam Stepper',
@@ -10,17 +11,14 @@ export const metadata = {
 export default function RootLayout({children}: {
     children: React.ReactNode
 }) {
-    const currentYear = new Date().getFullYear();
     return (
         <html lang="en">
-        <body className="min-h-screen flex flex-col font-sans text-gray-800">
+        <body className="min-h-screen flex flex-col">
         <Header/>
         <main className="flex-1">
             {children}
         </main>
-        <footer className="py-4 text-center text-sm text-gray-500">
-            ©{currentYear}
-        </footer>
+        <Footer/>
         </body>
         </html>
     );
